requirements.txt
Gradio
groq
python-dotenv